+++ packages/hooks/src/orderly/useExecutionReport.ts
@@ -5,21 +5,24 @@ export const useExecutionReport = (options?: {
   onMessage?: (data: any) => void;
 }) => {
   const ws = useWS();
-  const { data } = useSWRSubscription("executionreport", (_, { next }) => {
-    const unsubscribe = ws.privateSubscribe({
-      id: "executionreport",
-      event: "subscribe",
-      topic: "executionreport",
-      ts: Date.now(),
-    }, {
-      onMessage: (data: any) => {
-        //
-        options?.onMessage?.(data);
-        next(data);
+  return useSWRSubscription("executionreport", (_, { next }) => {
+    const unsubscribe = ws.privateSubscribe(
+      {
+        id: "executionreport",
+        event: "subscribe",
+        topic: "executionreport",
+        ts: Date.now(),
       },
-    });
+      {
+        onMessage: (data: any) => {
+          //
+          options?.onMessage?.(data);
+          next(null, data);
+        },
+      }
+    );
     return () => unsubscribe();
   });
 
-  return data;
+  // return data;
 };