+++ packages/react/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @orderly.network/react
 
+## 0.0.86
+
+### Patch Changes
+
+- @orderly.network/hooks@0.0.91
+- @orderly.network/components@0.0.88
+
+## 0.0.85
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/components@0.0.88
+  - @orderly.network/hooks@0.0.90
+
+## 0.0.84
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/components@0.0.87
+  - @orderly.network/hooks@0.0.89
+
+## 0.0.83
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/components@0.0.86
+  - @orderly.network/hooks@0.0.88
+
 ## 0.0.82
 
 ### Patch Changes